+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 ### Added
-- Added a new Traffic Ops endpoint to `GET` capacity and telemtry data for CDNi integration.
+- Added a new Traffic Ops endpoint to `GET` capacity and telemetry data for CDNi integration.
 
 ### Fixed
 - Update traffic_portal dependencies to mitigate `npm audit` issues.